requirements.txt
streamlit==1.40.0
google-generativeai
joblib
python-dotenv